requirements.txt
anthropic
python-dotenv
rich
openai
chromadb